% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetLEA}
\alias{GetLEA}
\title{A Function to Return DC LEA(Local Education Authority)/Sector Level Education Data}
\usage{
GetLEA("exhibit")
}
\description{
This function retrieves the data from the exhibits on LearnDC.org's LEA/Sector Report Cards and Profiles.
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/search/lea}
}
\examples{
lea_exhibit_names <- c("graduation","dccas","hqt_classes","staff_degree","mgp_scores","special_ed","enrollment"))
lea_grad <- GetLEA("graduation")
}
