% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetLEAExhibits}
\alias{GetLEAExhibits}
\title{A Function to Return School Exhibit Names}
\usage{
GetLEAExhibits()
}
\description{
This function retrieves the names of the exhibits from LearnDC's LEA/Sector Report Cards.
}
\examples{
GetLEAExhibits()
}
