% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetSchool}
\alias{GetSchool}
\title{A Function to Return DC School Level Data}
\usage{
GetSchool("exhibit")
}
\description{
This function retrieves the data from the exhibits on LearnDC.org's School Profiles, Report Cards, and Equity Reports.
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/search}
}
\examples{
GetSchool("graduation")
}
