% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetSchoolExhibits}
\alias{GetSchoolExhibits}
\title{A Function to Return School Exhibit Names}
\usage{
GetSchoolExhibits()
}
\description{
This function retrieves the names of the exhibits from LearnDC's School Profiles.
}
\examples{
GetSchoolExhibits()
}
