% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetSector}
\alias{GetSector}
\title{A Function to Return DC Sector (i.e. overall Public and Public Charter Schools) Level Education Data}
\usage{
GetSector("exhibit")
}
\description{
This function retrieves the data from the exhibits on LearnDC.org's Sector Report Cards and Profiles.
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/view?s=lea-0000#reportcard}
\url{http://learndc.org/schoolprofiles/view?s=lea-0001#reportcard}
}
\examples{
sector_exhibit_names <- c("graduation","dccas","hqt_classes","staff_degree","mgp_scores","special_ed","enrollment"))
sector_grad <- GetSector("graduation")
}
