% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetSectorExhibits}
\alias{GetSectorExhibits}
\title{A Function to Return Exhibit Names from LearnDC's Sector Report Cards}
\usage{
GetSectorExhibits()
}
\description{
This function retrieves the names of the exhibits from LearnDC.org's Sector Profiles and Report Cards.
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/view?s=lea-0000#reportcard}
\url{http://learndc.org/schoolprofiles/view?s=lea-0001#reportcard}
}
\examples{
GetSectorExhibits()
}
