% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetState}
\alias{GetState}
\title{A Function to Return DC/State Level Education Data}
\usage{
GetState("exhibit")
}
\description{
This function retrieves the data from the exhibits on LearnDC.org's DC/State Report Card.
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/view?s=dc#reportcard}
}
\examples{
GetState("graduation")
}
