% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetStateExhibits}
\alias{GetStateExhibits}
\title{A Function to Return School Exhibit Names}
\usage{
GetStateExhibits()
}
\description{
This function retrieves the names of the exhibits from LearnDC's DC State Report Card.
}
\author{
Benjamin Robinson, \email{Work:  benjamin.robinson@dc.gov,Personal:  benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/view?s=dc#reportcard}
}
\examples{
GetStateExhibits()
}
