% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLEA.R
\name{GetLEA}
\alias{GetLEA}
\title{A Function to Return DC LEA(Local Education Authority) Level Education Data}
\usage{
GetLEA("exhibit")
}
\arguments{
\item{exhibit}{character, exhibit name.  one of 
c("graduation","dccas","hqt_classes","staff_degree","mgp_scores",
"special_ed","enrollment","parcc")}
}
\value{
data frame
}
\description{
This function retrieves the data from the exhibits on 
LearnDC.org's LEA Report Cards and Profiles (does not include DC Public 
Schools or DC Public Charter Schools aggregate data, to retrieve that 
data, please use the GetSector() function.
}
\examples{
library(LearnDC)
lea_grad <- GetLEA("graduation")
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/search/lea}
}

