% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetLEAExhibits}
\alias{GetLEAExhibits}
\title{A Function to Return Exhibit Names from LearnDC's LEA Report Cards}
\usage{
GetLEAExhibits()
}
\description{
This function retrieves the names of the exhibits from LearnDC.org's LEA Profiles and Report Cards.
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/search/lea}
}
\examples{
GetLEAExhibits()
}
