% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSchool.R
\name{GetSchool}
\alias{GetSchool}
\title{A Function to Return DC School Level Data}
\usage{
GetSchool("exhibit")
}
\arguments{
\item{exhibit}{character, exhibit name.  one of 
c("graduation","dccas","attendance","hqt_classes","staff_degree",
"mgp_scores","special_ed","enrollment","suspensions","expulsions",
"enrollment_equity","accountability","accountability_classification")}
}
\value{
data frame
}
\description{
This function retrieves the data from the exhibits on 
LearnDC.org's School Profiles, Report Cards, and Equity Reports.
}
\examples{
school_grad <- GetSchool("graduation")
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/search}
}

