% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetSchoolExhibits}
\alias{GetSchoolExhibits}
\title{A Function to Return School Exhibit Names}
\usage{
GetSchoolExhibits()
}
\description{
This function retrieves the names of the exhibits from LearnDC's School Profiles.
}
\author{
Benjamin Robinson, \email{Work:  benjamin.robinson@dc.gov,Personal:  benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/search}
}
\examples{
GetSchoolExhibits()
}
