% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetSchoolExhibits}
\alias{GetSchoolExhibits}
\title{A Function to Return Exhibit Names from LearnDC's School Profiles}
\usage{
GetSchoolExhibits()
}
\description{
This function retrieves the names of the exhibits from LearnDC.org's School Profiles, Report Cards, and Equity Reports.
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/search}
}
\examples{
GetSchoolExhibits()
}
