% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSector.R
\name{GetSector}
\alias{GetSector}
\title{A Function to Return DC Sector (i.e. overall Public and 
Public Charter Schools) Level Education Data}
\usage{
GetSector("exhibit")
}
\arguments{
\item{exhibit}{character, exhibit name.  one of 
c("graduation","dccas","hqt_classes","staff_degree","mgp_scores",
"special_ed","enrollment")}
}
\value{
data frame
}
\description{
This function retrieves the data from the exhibits on 
LearnDC.org's Sector Report Cards and Profiles.
}
\examples{
sector_grad <- GetSector("graduation")
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/view?s=lea-0000#reportcard}
\url{http://learndc.org/schoolprofiles/view?s=lea-0001#reportcard}
}

