% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetState.R
\name{GetState}
\alias{GetState}
\title{A Function to Return DC/State Level Education Data}
\usage{
GetState("exhibit")
}
\arguments{
\item{exhibit}{character, exhibit name.  one of 
c("graduation","dccas","attendance","naep_results","hqt_classes",
"staff_degree","mgp_scores","ell","special_ed","enrollment",
"suspensions","expulsions","enrollment_equity","accountability",
"amo_targets","expulsions")}
}
\value{
data frame
}
\description{
This function retrieves the data from the exhibits on 
LearnDC.org"s DC/State Report Card.
}
\examples{
library(LearnDC)
state_grad <- GetState("graduation")
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/view?s=dc#reportcard}
}

