% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetState}
\alias{GetState}
\title{A Function to Return DC/State Level Education Data}
\usage{
GetState("exhibit")
}
\description{
This function retrieves the data from the exhibits on LearnDC.org"s DC/State Report Card.
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/view?s=dc#reportcard}
}
\examples{
state_exhibit_names <- c("graduation","dccas","attendance","naep_results","hqt_classes","staff_degree","mgp_scores","ell","special_ed","enrollment","suspensions","expulsions","enrollment_equity","accountability","amo_targets","expulsions")
state_grad <- GetState("graduation")
}
