% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetStateExhibits}
\alias{GetStateExhibits}
\title{A Function to Return Exhibit Names from LearnDC's DC State Report Card}
\usage{
GetStateExhibits()
}
\description{
This function retrieves the names of the exhibits from LearnDC.org's DC State Report Card, Profile, and Equity Report.
}
\author{
Benjamin Robinson, \email{benj.robinson2@gmail.com}
}
\references{
\url{http://learndc.org/schoolprofiles/view?s=dc#reportcard}
}
\examples{
GetStateExhibits()
}
