% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{GetStateExhibits}
\alias{GetStateExhibits}
\title{A Function to Return School Exhibit Names}
\usage{
GetStateExhibits()
}
\description{
This function retrieves the names of the exhibits from LearnDC's DC State Report Card.
}
\examples{
GetStateExhibits()
}
